move space station generator to playground
